# run_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Belajar Laravel",
        "description": "Dokumentasi Api Belajar Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/belajar-laravel/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register User",
                "description": "Register New User",
                "operationId": "1ae02d0cc5ded760f9bf5a644c85cef7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jhon Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jhondoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register New User Data"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/belajar-laravel/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login (Authentitication)",
                "operationId": "login",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Enter your email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Enter password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/belajar-laravel/api/cicilan": {
            "get": {
                "tags": [
                    "Cicilan"
                ],
                "summary": "Get Data Cicilan",
                "operationId": "e019682c0addf691d2441419fe333d5f",
                "responses": {
                    "200": {
                        "description": "Get Data List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/belajar-laravel/api/cicilan/{id}": {
            "get": {
                "tags": [
                    "Cicilan"
                ],
                "summary": "Get Data by Id",
                "operationId": "f1bceecc30c6826c2b82c57aa3b80713",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inputkan Id Hutang",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Data List by Id"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/belajar-laravel/api/hutang": {
            "get": {
                "tags": [
                    "Hutang"
                ],
                "summary": "Get Data Hutang",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Get Data List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hutang"
                ],
                "summary": "Store Data Hutang",
                "operationId": "store",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "description": "Inputkan Nama",
                                        "type": "string"
                                    },
                                    "jenis_kelamin": {
                                        "description": "Pilih Jenis Kelamin",
                                        "type": "string",
                                        "enum": [
                                            "Laki-laki",
                                            "Perempuan"
                                        ]
                                    },
                                    "alamat": {
                                        "description": "Inputkan Alamat",
                                        "type": "string"
                                    },
                                    "tanggal_hutang": {
                                        "description": "Inputkan Tanggal Hutang",
                                        "type": "string",
                                        "example": "12-12-2012"
                                    },
                                    "hutang": {
                                        "description": "Inputkan Jumlah Hutang",
                                        "type": "integer"
                                    },
                                    "cicilan_id": {
                                        "description": "Inputkan Cicilan Id",
                                        "type": "integer"
                                    },
                                    "jaminan": {
                                        "description": "Upload Jaminan",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/belajar-laravel/api/hutang/{id}": {
            "get": {
                "tags": [
                    "Hutang"
                ],
                "summary": "Get Data by Id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inputkan Id Hutang",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Data List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hutang"
                ],
                "summary": "Update Data Hutang",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inputkan Id Hutang",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "JANGAN DIUBAH",
                                        "example": "PUT"
                                    },
                                    "nama": {
                                        "description": "Inputkan Nama",
                                        "type": "string"
                                    },
                                    "jenis_kelamin": {
                                        "description": "Pilih Jenis Kelamin",
                                        "type": "string",
                                        "enum": [
                                            "Laki-laki",
                                            "Perempuan"
                                        ]
                                    },
                                    "alamat": {
                                        "description": "Inputkan Alamat",
                                        "type": "string"
                                    },
                                    "tanggal_hutang": {
                                        "description": "Inputkan Tanggal Hutang",
                                        "type": "string",
                                        "example": "12-12-2012"
                                    },
                                    "hutang": {
                                        "description": "Inputkan Jumlah Hutang",
                                        "type": "integer"
                                    },
                                    "cicilan_id": {
                                        "description": "Pilih Jenis Cicilan",
                                        "type": "integer"
                                    },
                                    "jaminan": {
                                        "description": "Upload Jaminan",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hutang"
                ],
                "summary": "Destroy Data Hutang",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inputkan Id Hutang",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Data List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "sanctum",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}